requirements.txt
torch
transformers
pytorch-crf